Fix adc bug
The symptom of this bug was a fraction of the value in ADC0 was being
added to all other values. It seems to be related to the fact that when
initializing an ADC read we would first clear the low bits in ADCMUX
before writing the new channel index. Having a 0 in the channel
selection bits in between clearing them and writing the new value seems
to result in some amount of the voltage at ADC0 contributing to the next
value read.